Move command to own class
