+++ modules/dkan_datastore/drush.services.yml
@@ -26,3 +26,9 @@ services:
       - '@dkan.datastore.logger_channel'
     tags:
       - { name: drush.command }
+  dkan_datastore.degraded_mode.commands:
+    class: \Drupal\dkan_datastore\Commands\DegradedModeCommands
+    arguments:
+      - '@state'
+    tags:
+      - { name: drush.command }
