+++ modules/dkan_datastore/src/Drush.php
@@ -370,31 +370,4 @@ class Drush extends DrushCommands {
     return DrushCommands::EXIT_FAILURE;
   }
 
-  /**
-   * Enable or disable degraded datastore query mode.
-   *
-   * @param string|null $state
-   *   Optional state: 1 or 0. If omitted, current status shown.
-   *
-   * @command dkan:datastore:degraded-mode
-   */
-  public function degradedMode(?string $state = NULL) {
-    $current = (bool) $this->state->get('dkan_datastore.degraded_performance', FALSE);
-
-    if ($state === NULL) {
-      $this->output()->writeln('Datastore degraded performance mode: ' . ($current ? 'enabled' : 'disabled'));
-      return DrushCommands::EXIT_SUCCESS;
-    }
-
-    if ($state !== '1' && $state !== '0') {
-      $this->output()->writeln('Invalid value. Use 1 or 0.');
-      return DrushCommands::EXIT_FAILURE;
-    }
-
-    $value = $state === '1';
-    $this->state->set('dkan_datastore.degraded_performance', $value);
-    $this->output()->writeln('Datastore degraded performance mode ' . ($value ? 'enabled' : 'disabled') . '.');
-    return DrushCommands::EXIT_SUCCESS;
-  }
-
 }
